requirements.txt
python-telegram-bot>=21.6,<22
requests>=2.32.3,<3
python-dotenv>=1.0.1,<2
icalendar>=5.0.12,<6
pydantic>=2.8.2,<3
pydantic-settings>=2.6.0,<3
python-slugify>=8.0.4,<9
beautifulsoup4>=4.12.3,<5
pytest>=8.3.2,<9
